requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
langchain-text-splitters>=0.0.1

# Vector Store
faiss-cpu>=1.7.4



# Document Loaders
pypdf>=3.17.0
docx2txt>=0.8
openpyxl>=3.1.0
unstructured>=0.10.0

# Hybrid Search (BM25)
rank-bm25>=0.2.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0

# Optional: ChromaDB (alternative to FAISS)
# chromadb>=0.4.0